requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anthropic==0.31.2
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
billiard==4.2.0
boto3==1.35.0
botocore==1.35.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cohere==5.2.2
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
cron-descriptor==1.4.3
cryptography==42.0.5
cycler==0.12.1
daphne==4.1.2
dataclasses-json==0.6.4
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
Django==5.0.4
django-celery-results==2.2.0
django-cors-headers==4.3.1
django-timezone-field==7.0
djangorestframework==3.15.1
et-xmlfile==1.1.0
eval_type_backport==0.2.0
fastavro==1.9.4
fasttext-wheel==0.9.2
filelock==3.13.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
grpcio==1.62.1
grpcio-health-checking==1.62.1
grpcio-tools==1.62.1
gunicorn==21.2.0
h11==0.14.0
html2markdown==0.1.7
htmltabletomd==1.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
humanize==4.9.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
kombu==5.3.6
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.40
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langsmith==0.1.40
llama-index==0.10.44
llama-index-agent-openai==0.2.7
llama-index-cli==0.1.12
llama-index-core==0.10.44
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.22
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.25
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.4
llamaindex-py-client==0.1.19
loader==2017.9.11
Logbook==1.7.0.post0
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openai==1.16.2
openpyxl==3.1.5
orjson==3.10.0
packaging==23.2
pandas==2.2.2
pendulum==3.0.0
pillow==10.3.0
pptx-replace==0.1.5
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==6.0.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.13.4
pycparser==2.22
pydantic
Pygments==2.18.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypdf==4.2.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-docx-replace==0.4.4
python-dotenv==1.0.1
python-pptx==1.0.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
regex==2023.12.25
requests==2.31.0
rich==13.7.1
s3transfer==0.10.2
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.4.4
striprtf==0.0.26
tabulate==0.9.0
tenacity==8.2.3
termcc==2019.8.2
terminaltables==3.1.10
tiktoken==0.6.0
time-machine==2.14.1
together==1.2.5
tokenizers==0.15.2
tornado==6.4
tqdm==4.66.2
Twisted==24.3.0
txaio==23.1.1
typer==0.12.3
types-requests==2.31.0.20240403
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
validators==0.22.0
vine==5.1.0
wcwidth==0.2.13
weaviate-client==4.5.5
whitenoise==6.7.0
wrapt==1.16.0
XlsxWriter==3.2.0
yarl==1.9.4
zope.interface==6.4.post2
